Fix nits and add handle resolver host flag for Bigsky
